requirements.txt
fastapi==0.116.0
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.4
pydantic-settings==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
pypdf==5.4.0
python-docx==1.1.2
openpyxl==3.1.5
azure-cosmos==4.9.0
azure-identity==1.17.1